==> 01_growth_rate.Rmd <==

# Growth Rate {#growth-rate}

## Query 1 - Adaptation

Applied on Scopus. 

```

9,873 document results

TITLE-ABS-KEY ( ( "Climat* Change"  OR  "global warming" )  AND  
( agricult*  OR  crop  OR  cultiva*  OR  farm*  OR  tillage )  AND  
adapt*  AND NOT  mitig* )  AND  
( LIMIT-TO ( DOCTYPE ,  "ar" )  OR  LIMIT-TO ( DOCTYPE ,  "re" ) ) 
```

Downloaded in March 2022.

```{r eval = F, echo = F, warning = FALSE, error = TRUE, tidy = FALSE, message = FALSE}

tic()

M <- convert2df(file = c("bibs/adaptation_scopus1.bib", 
                         "bibs/adaptation_scopus2.bib", 
                         "bibs/adaptation_scopus3.bib", 
                         "bibs/adaptation_scopus4.bib", 
                         "bibs/adaptation_scopus5.bib"), 
                dbsource = "scopus", format = "bibtex")


rio::export(M, 'rawfiles/M_adaptation.rds')
rio::export(M, 'rawfiles/M_adaptation.csv')

toc()

```

[Download csv file.](rawfiles/rawfiles/M_adaptation.csv)

## Growth rate

```{r eval = TRUE, echo = F, warning = FALSE, error = TRUE, tidy = FALSE, message = FALSE}

M <- import('rawfiles/M_adaptation.rds')

M %>>%
    count(PY, sort = F, name = 'Papers') %>>%
    (~ d0) %>>%
    dplyr::filter(PY %in% c(1990:2021)) %>>% 
    dplyr::arrange(PY) %>>% 
    dplyr::mutate(trend = 1:n()) %>>% 
    (. -> d)

d$lnp <- log(d$Papers)

# ajustar parametros via mqo 
m1 <- lm(lnp ~ trend, data = d)
# summary(m1)

beta0 <- m1$coefficients[[1]]
beta1 <- m1$coefficients[[2]]

# modelo não linear
# 2000 é o primeiro ano da série
m2 <- nls(Papers ~ b0 * exp(b1 * (PY - 1990)), start = list(b0 = beta0, b1 = beta1), data = d)
# summary(m2)

# publications estimado
d$predicted <- 8.973890 * exp(0.160793 * (d$PY - 1990))

# taxa de crescimento anual é de 
# (exp(0.160793) - 1) * 100

# o período necessário para dobrar o tempo é 
# log(2) / 0.160793

d %>>% 
    mutate(Year = PY) %>>% 
    mutate(predicted = round(predicted, 0)) %>>% 
    (. -> d)

periodo <- 1990:2024
predicted <- tibble::tibble(PY = periodo, Predicted = round(8.973890 * exp(0.160793 * (periodo - 1990)), 0)) 

dplyr::full_join(d0 |> dplyr::filter(PY <= 2021), predicted) |>
    dplyr::rename(Year = PY) |>
    dplyr::filter(Year >= 1990) ->
    d2

highcharter::hchart(d2, "column", hcaes(x = Year, y = Papers), name = "Publications", showInLegend = TRUE) %>>%
   highcharter::hc_add_series(d2, "line", hcaes(x = Year, y = Predicted), name = "Predicted", showInLegend = TRUE) %>>% 
   highcharter::hc_add_theme(hc_theme_google()) %>>%
   highcharter::hc_navigator(enabled = TRUE)  %>>% 
   highcharter::hc_exporting(enabled = TRUE, filename = 'groups_growth') %>>%
   highcharter::hc_xAxis(plotBands = list(list(from = 2021, to = 2021, color = "#330000")))
```

Analysis 1990-2021
<ul>
  <li> Growth Rate 17.4% </li>
  <li> Doubling time  4.1 Years </li>
</ul>


## Query 2 - Mitigation

```
TITLE-ABS-KEY ( ( "Climat* Change"  OR  "global warming" )  AND  
( agricult*  OR  crop  OR  cultiva*  OR  farm*  OR  tillage )  AND  mitig*  AND NOT adapt*)  AND  
( LIMIT-TO ( DOCTYPE ,  "ar" ) )
```

```{r eval = F, echo = F, warning = FALSE, error = TRUE, tidy = FALSE, message = FALSE}

M <- convert2df(file = c("bibs/mitigation_scopus1.bib", 
                         "bibs/mitigation_scopus2.bib", 
                         "bibs/mitigation_scopus3.bib"), 
                dbsource = "scopus", format = "bibtex")


rio::export(M, 'rawfiles/M_mitigation.rds')
rio::export(M, 'rawfiles/M_mitigation.csv')
```

[Download csv file.](rawfiles/rawfiles/M_mitigation.csv)

## Growth rate

```{r eval = TRUE, echo = F, warning = FALSE, error = TRUE, tidy = FALSE, message = FALSE}

M <- rio::import('rawfiles/M_mitigation.rds')

M %>>%
    count(PY, sort = F, name = 'Papers') %>>%
    (~ d0) %>>%
    dplyr::filter(PY %in% c(1990:2021)) %>>% 
    dplyr::arrange(PY) %>>% 
    dplyr::mutate(trend = 1:n()) %>>% 
    (. -> d)

d$lnp <- log(d$Papers)

# ajustar parametros via mqo 
m1 <- lm(lnp ~ trend, data = d)
# summary(m1)

beta0 <- m1$coefficients[[1]]
beta1 <- m1$coefficients[[2]]

# modelo não linear
# 2000 é o primeiro ano da série
m2 <- nls(Papers ~ b0 * exp(b1 * (PY - 1990)), start = list(b0 = beta0, b1 = beta1), data = d)
# summary(m2)

# publications estimado
d$predicted <- 3.725894 * exp(0.174929 * (d$PY - 1990))

# taxa de crescimento anual é de 
# (exp(0.174929) - 1) * 100

# o período necessário para dobrar o tempo é 
# log(2) / 0.174929

d %>>% 
    mutate(Year = PY) %>>% 
    mutate(predicted = round(predicted, 0)) %>>% 
    (. -> d)

periodo <- 1990:2024
predicted <- tibble::tibble(PY = periodo, Predicted = round(3.725894 * exp(0.174929 * (periodo - 1990)), 0)) 

dplyr::full_join(d0 |> dplyr::filter(PY <= 2021), predicted) |>
    dplyr::rename(Year = PY) |>
    dplyr::filter(Year >= 1990) ->
    d2

highcharter::hchart(d2, "column", hcaes(x = Year, y = Papers), name = "Publications", showInLegend = TRUE) %>>%
   highcharter::hc_add_series(d2, "line", hcaes(x = Year, y = Predicted), name = "Predicted", showInLegend = TRUE) %>>% 
   highcharter::hc_add_theme(hc_theme_google()) %>>%
   highcharter::hc_navigator(enabled = TRUE)  %>>% 
   highcharter::hc_exporting(enabled = TRUE, filename = 'groups_growth') %>>%
   highcharter::hc_xAxis(plotBands = list(list(from = 2021, to = 2021, color = "#330000")))
```

Analysis 1990-2021
<ul>
  <li> Growth Rate 19.1% </li>
  <li> Doubling time  3.9 Years </li>
</ul>


## Scopus

```{r eval = TRUE, echo = F, warning = FALSE, error = TRUE, tidy = FALSE, message = FALSE}

# alterar arquivos csv na mão para facilitar a vida
d0 <- read.table('rawfiles/Scopus-Query1-Analyze-Year.csv', header = T, sep = ',') 

d0 |>
    dplyr::mutate(Papers = Papers / 100000) |> 
    dplyr::filter(PY %in% c(1990:2021)) %>>% 
    dplyr::arrange(PY) %>>% 
    dplyr::mutate(trend = 1:n()) %>>% 
    (. -> d)

d$lnp <- log(d$Papers)

# ajustar parametros via mqo 
m1 <- lm(lnp ~ trend, data = d)
# summary(m1)

beta0 <- m1$coefficients[[1]]
beta1 <- m1$coefficients[[2]]

# modelo não linear
# 2000 é o primeiro ano da série
m2 <- nls(Papers ~ b0 * exp(b1 * (PY - 1990)), start = list(b0 = beta0, b1 = beta1), data = d)
# summary(m2)

# publications estimado
d$predicted <- 6.488682 * exp(0.046916 * (d$PY - 1990))

# taxa de crescimento anual é de 
# (exp(0.046916) - 1) * 100

# o período necessário para dobrar o tempo é 
# log(2) / 0.046916

d %>>% 
    mutate(Year = PY) %>>% 
    mutate(predicted = round(predicted, 0)) %>>% 
    (. -> d)

periodo <- 1990:2024
predicted <- tibble::tibble(PY = periodo, Predicted = round(6.488682 * exp(0.046916 * (periodo - 1990)), 0)) 

dplyr::full_join(d0 |> dplyr::filter(PY <= 2021), predicted) |>
    dplyr::arrange(PY) |> 
    dplyr::rename(Year = PY) |>
    dplyr::filter(Year >= 1990) |>
    dplyr::mutate(Papers = Papers / 100000) -> 
    d2

highcharter::hchart(d2, "column", hcaes(x = Year, y = Papers), name = "Publications", showInLegend = TRUE) %>>%
   highcharter::hc_add_series(d2, "line", hcaes(x = Year, y = Predicted), name = "Predicted", showInLegend = TRUE) %>>% 
   highcharter::hc_add_theme(hc_theme_google()) %>>%
   highcharter::hc_navigator(enabled = TRUE)  %>>% 
   highcharter::hc_exporting(enabled = TRUE, filename = 'groups_growth') %>>%
   highcharter::hc_xAxis(plotBands = list(list(from = 2021, to = 2021, color = "#330000")))

```

Analysis 1990-2020
<ul>
  <li> Growth Rate 4.8% </li>
  <li> Doubling time  14.7 Years </li>
</ul>
